chore: Update WebSocket connection in client and email template in server
This commit updates the WebSocket connection in the client code to establish a connection using the provided token. It also removes the commented out code for connecting the socket.

In the server code, the email template for verification emails has been updated. The `html` property has been replaced with a call to the `pug.renderFile` function, passing the necessary data to render the verification code email template.

These changes improve the functionality and maintainability of the application.